member에 토큰 저장 -> 2차인증처리

+++ src/main/java/com/lion/spring_paymodule/app/util/JwtProvider.java
@@ -27,7 +27,7 @@ public class JwtProvider {
         return secretKey;
     }
 
-    public String generateAccessToken(Map<String, Object> claims, int seconds) {
+    public String generateAccessToken(Map<String, Object> claims, long seconds) {
         long now = new Date().getTime();
         Date accessTokenExpiresIn = new Date(now + 1000L * seconds);
